# server-app/src/main/resources/static/openapi.yaml
openapi: 3.1.1
info:
  title: BancoSimple API
  version: 1.0.1
  description: Documentación combinada de los módulos de la API
paths:
  /api/activity-logs:
    get:
      summary: Obtener todos los registros de actividad
      description: Devuelve todos los logs de acciones realizadas por los usuarios
        en el sistema.
      tags:
      - Logs de Actividad
      responses:
        '200':
          description: Lista de logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ActivityLogDTO'
    post:
      summary: Registrar nueva actividad
      description: Registra una acción realizada por un usuario en el sistema.
      tags:
      - Logs de Actividad
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActivityLogDTO'
      responses:
        '200':
          description: Log registrado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityLogDTO'
  /api/activity-logs/{id}:
    get:
      summary: Obtener log de actividad por ID
      description: Devuelve los detalles de un registro de actividad específico.
      tags:
      - Logs de Actividad
      parameters:
      - name: id
        in: path
        required: true
        description: ID del log
        schema:
          type: integer
      responses:
        '200':
          description: Log encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityLogDTO'
    delete:
      summary: Eliminar log de actividad
      description: Elimina un log del sistema.
      tags:
      - Logs de Actividad
      parameters:
      - name: id
        in: path
        required: true
        description: ID del log
        schema:
          type: integer
      responses:
        '204':
          description: Log eliminado correctamente
...
  /api/frequent-accounts:
    get:
      summary: Obtener todas las cuentas frecuentes
      description: Devuelve la lista de cuentas frecuentes registradas por los usuarios.
      tags:
      - Cuentas Frecuentes
      responses:
        '200':
          description: Lista de cuentas frecuentes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FrequentAccountDTO'
    post:
      summary: Agregar cuenta frecuente
      description: Agrega una nueva cuenta frecuente asociada a un usuario.
      tags:
      - Cuentas Frecuentes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FrequentAccountDTO'
      responses:
        '200':
          description: Cuenta frecuente agregada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrequentAccountDTO'
  /api/frequent-accounts/{id}:
    get:
      summary: Obtener cuenta frecuente por ID
      description: Devuelve los detalles de una cuenta frecuente específica.
      tags:
      - Cuentas Frecuentes
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
        description: ID de la cuenta frecuente
      responses:
        '200':
          description: Cuenta frecuente encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrequentAccountDTO'
    put:
      summary: Actualizar cuenta frecuente
      description: Actualiza la información de una cuenta frecuente.
      tags:
      - Cuentas Frecuentes
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
        description: ID de la cuenta frecuente
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FrequentAccountDTO'
      responses:
        '200':
          description: Cuenta frecuente actualizada correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrequentAccountDTO'
    delete:
      summary: Eliminar cuenta frecuente
      description: Elimina una cuenta frecuente del sistema.
      tags:
      - Cuentas Frecuentes
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
        description: ID de la cuenta frecuente
      responses:
        '204':
          description: Cuenta frecuente eliminada correctamente
...
  /api/products:
    get:
      summary: Obtener todos los productos
      description: Devuelve una lista de todos los productos registrados para todos
        los usuarios.
      tags:
      - Productos
      responses:
        '200':
          description: Lista de productos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductDTO'
    post:
      summary: Crear un nuevo producto
      description: Crea un producto nuevo para un usuario específico.
      tags:
      - Productos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductDTO'
      responses:
        '200':
          description: Producto creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDTO'
  /api/products/{id}:
    get:
      summary: Obtener producto por ID
      description: Devuelve los detalles de un producto específico.
      tags:
      - Productos
      parameters:
      - name: id
        in: path
        required: true
        description: ID del producto
        schema:
          type: integer
      responses:
        '200':
          description: Producto encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDTO'
    put:
      summary: Actualizar producto
      description: Actualiza los datos de un producto existente.
      tags:
      - Productos
      parameters:
      - name: id
        in: path
        required: true
        description: ID del producto
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductDTO'
      responses:
        '200':
          description: Producto actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDTO'
    delete:
      summary: Eliminar producto
      description: Elimina un producto por su ID.
      tags:
      - Productos
      parameters:
      - name: id
        in: path
        required: true
        description: ID del producto
        schema:
          type: integer
      responses:
        '204':
          description: Producto eliminado correctamente
  /api/products/request:
    post:
      summary: Solicitar producto
      description: Permite que un usuario solicite un nuevo producto, sin crearlo
        directamente.
      tags:
      - Productos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductDTO'
      responses:
        '200':
          description: Solicitud procesada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDTO'
  /api/products/{id}/status:
    patch:
      summary: Actualizar estado del producto
      description: Cambia el estado de un producto existente.
      tags:
      - Productos
      parameters:
      - name: id
        in: path
        required: true
        description: ID del producto
        schema:
          type: integer
      - name: estadoId
        in: query
        required: true
        description: Nuevo ID del estado
        schema:
          type: integer
      responses:
        '200':
          description: Estado actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDTO'
...
  /api/users:
    get:
      summary: Obtener todos los usuarios
      description: Devuelve una lista de todos los usuarios registrados en el sistema.
      tags:
      - Usuarios
      responses:
        '200':
          description: Lista de usuarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDTO'
    post:
      summary: Crear un nuevo usuario
      description: Registra un nuevo usuario en el sistema.
      tags:
      - Usuarios
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDTO'
      responses:
        '200':
          description: Usuario creado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
  /api/users/{id}:
    get:
      summary: Obtener usuario por ID
      description: Obtiene los datos de un usuario específico.
      tags:
      - Usuarios
      parameters:
      - name: id
        in: path
        required: true
        description: ID del usuario
        schema:
          type: integer
      responses:
        '200':
          description: Usuario encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
    put:
      summary: Actualizar usuario
      description: Actualiza la información de un usuario existente.
      tags:
      - Usuarios
      parameters:
      - name: id
        in: path
        required: true
        description: ID del usuario
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDTO'
      responses:
        '200':
          description: Usuario actualizado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
    delete:
      summary: Eliminar usuario
      description: Elimina un usuario por su ID.
      tags:
      - Usuarios
      parameters:
      - name: id
        in: path
        required: true
        description: ID del usuario
        schema:
          type: integer
      responses:
        '204':
          description: Usuario eliminado correctamente
...
components:
  schemas:
    ActivityLogDTO:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        action:
          type: string
          example: LOGIN
        sourceIp:
          type: string
          example: 192.168.1.1
        date:
          type: string
          format: date-time
        details:
          type: string
          example: Inicio de sesión exitoso desde cliente web
...
    FrequentAccountDTO:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
          enum:
          - INTERNAL
          - EXTERNAL
        favoriteProductId:
          type: integer
          nullable: true
        externalBankName:
          type: string
          enum:
          - BANCO_DE_CHILE
          - BANCO_ESTADO
          - SANTANDER
          - BCI
          - ITAU
          - FALABELLA
          - CENCOSUD
          - SECURITY
          - BICE
          - SCOTIABANK
          - HSBC
          - CORPBANCA
          nullable: true
        externalAccountNumber:
          type: string
          nullable: true
        externalHolderName:
          type: string
          nullable: true
        alias:
          type: string
          example: Cuenta de mamá
        addedDate:
          type: string
          format: date-time
        active:
          type: boolean
...
    ProductDTO:
      type: object
      properties:
        id:
          type: integer
        usuarioId:
          type: integer
        tipoProducto:
          type: string
        numeroProducto:
          type: string
        alias:
          type: string
        moneda:
          type: string
        saldo:
          type: number
          format: double
        limiteCredito:
          type: number
          format: double
        estadoId:
          type: integer
        creadoEn:
          type: string
          format: date-time
...
    UserDTO:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nombres:
          type: string
          example: Juan
        apellidos:
          type: string
          example: Pérez
        correo:
          type: string
          example: juan.perez@example.com
        telefono:
          type: string
          example: '+56912345678'
        documentoTipo:
          type: string
          example: RUT
        documentoNumero:
          type: string
          example: 12.345.678-9
        tipoUsuario:
          type: string
          example: cliente
        estadoId:
          type: integer
          example: 1
        fechaNacimiento:
          type: string
          format: date
        fechaRegistro:
          type: string
          format: date-time
...